Extract rasp NetworkConnection logic to a module to be able to reuse it in different advices
