+++ dd-java-agent/instrumentation/jackson-core/jackson-core-1/src/main/java/datadog/trace/instrumentation/jackson/codehouse/core/RaspJson1FactoryInstrumentation.java
@@ -7,10 +7,13 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
+import datadog.trace.api.Config;
+import datadog.trace.instrumentation.appsec.rasp.modules.NetworkConnectionModule;
 import datadog.trace.instrumentation.appsec.utils.InstrumentationLogger;
 import java.io.InputStream;
 import java.io.Reader;
 import java.net.URL;
+import net.bytebuddy.asm.Advice;
 
 @AutoService(InstrumenterModule.class)
 public class RaspJson1FactoryInstrumentation extends InstrumenterModule.AppSec
@@ -22,7 +25,9 @@ public class RaspJson1FactoryInstrumentation extends InstrumenterModule.AppSec
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {InstrumentationLogger.class.getName()};
+    return new String[] {
+      InstrumentationLogger.class.getName(), NetworkConnectionModule.class.getName()
+    };
   }
 
   @Override
@@ -36,11 +41,26 @@ public class RaspJson1FactoryInstrumentation extends InstrumenterModule.AppSec
                     .or(takesArguments(Reader.class))
                     .or(takesArguments(URL.class))
                     .or(takesArguments(byte[].class))),
-        "datadog.trace.instrumentation.appsec.rasp.advices.NetworkConnectionRaspAdvice");
+        RaspJson1FactoryInstrumentation.class.getName() + "$NetworkConnectionRaspAdvice");
   }
 
   @Override
   public String instrumentedType() {
     return "org.codehaus.jackson.JsonFactory";
+  }
+
+  public static class NetworkConnectionRaspAdvice {
+
+    @Advice.OnMethodExit()
+    public static void onExit(@Advice.Argument(0) final Object input) {
+
+      if (!Config.get().isAppSecRaspEnabled()) {
+        return;
+      }
+      if (input == null || !(input instanceof URL)) {
+        return;
+      }
+      NetworkConnectionModule.INSTANCE.onNetworkConnection(input);
+    }
   }
 }
