+++ dd-java-agent/instrumentation/jackson-core/src/main/java/datadog/trace/instrumentation/jackson/core/RaspJson2FactoryInstrumentation.java
@@ -8,10 +8,13 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
+import datadog.trace.api.Config;
+import datadog.trace.instrumentation.appsec.rasp.modules.NetworkConnectionModule;
 import datadog.trace.instrumentation.appsec.utils.InstrumentationLogger;
 import java.io.InputStream;
 import java.io.Reader;
 import java.net.URL;
+import net.bytebuddy.asm.Advice;
 
 @AutoService(InstrumenterModule.class)
 public class RaspJson2FactoryInstrumentation extends InstrumenterModule.AppSec
@@ -23,7 +26,9 @@ public class RaspJson2FactoryInstrumentation extends InstrumenterModule.AppSec
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {InstrumentationLogger.class.getName()};
+    return new String[] {
+      InstrumentationLogger.class.getName(), NetworkConnectionModule.class.getName()
+    };
   }
 
   @Override
@@ -39,11 +44,26 @@ public class RaspJson2FactoryInstrumentation extends InstrumenterModule.AppSec
                             .or(takesArguments(Reader.class))
                             .or(takesArguments(URL.class))
                             .or(takesArguments(byte[].class)))),
-        "datadog.trace.instrumentation.appsec.rasp.advices.NetworkConnectionRaspAdvice");
+        RaspJson2FactoryInstrumentation.class.getName() + "$NetworkConnectionRaspAdvice");
   }
 
   @Override
   public String instrumentedType() {
     return "com.fasterxml.jackson.core.JsonFactory";
+  }
+
+  public static class NetworkConnectionRaspAdvice {
+
+    @Advice.OnMethodExit()
+    public static void onExit(@Advice.Argument(0) final Object input) {
+
+      if (!Config.get().isAppSecRaspEnabled()) {
+        return;
+      }
+      if (input == null || !(input instanceof URL)) {
+        return;
+      }
+      NetworkConnectionModule.INSTANCE.onNetworkConnection(input);
+    }
   }
 }
